+++ CHANGELOG.rst
@@ -15,6 +15,12 @@ Quetzal version numbers follow `semantic versioning <http://semver.org>`_.
 * Added Let's encrypt SSL certificate management.
 * Added deployment on GCP documentation.
 * Added Sphinx documentation structure.
+* Added file delete endpoint.
+* Added file state in base metadata.
+* Added a simple/naive workspace commit conflict detection.
+* Added global queries.
+* Added support for temporary files. They are uploaded on the users' bucket.
+* Added path as query parameter when uploading files.
 
 0.1.0 (2019-03-05)
 ------------------